requirements.txt
fastapi
uvicorn
pandas
numpy
sqlalchemy
psycopg2-binary
sentence-transformers
pgvector
python-dotenv
